Write cache was integrated

+++ core/src/test/java/com/orientechnologies/orient/core/storage/cache/local/twoq/ReadWriteDiskCacheTest.java
@@ -4,9 +4,8 @@ import com.orientechnologies.common.collection.closabledictionary.OClosableLinke
 import com.orientechnologies.common.directmemory.OByteBufferPool;
 import com.orientechnologies.common.serialization.types.OIntegerSerializer;
 import com.orientechnologies.common.serialization.types.OLongSerializer;
+import com.orientechnologies.common.serialization.types.OStringSerializer;
 import com.orientechnologies.common.types.OModifiableBoolean;
-import com.orientechnologies.orient.core.Orient;
-import com.orientechnologies.orient.core.config.OContextConfiguration;
 import com.orientechnologies.orient.core.config.OGlobalConfiguration;
 import com.orientechnologies.orient.core.exception.OAllCacheEntriesAreUsedException;
 import com.orientechnologies.orient.core.storage.OChecksumMode;
@@ -15,7 +14,6 @@ import com.orientechnologies.orient.core.storage.cache.OCacheEntryImpl;
 import com.orientechnologies.orient.core.storage.cache.OCachePointer;
 import com.orientechnologies.orient.core.storage.cache.local.OWOWCache;
 import com.orientechnologies.orient.core.storage.fs.OFileClassic;
-import com.orientechnologies.orient.core.storage.impl.local.paginated.OLocalPaginatedStorage;
 import com.orientechnologies.orient.core.storage.impl.local.paginated.base.ODurablePage;
 import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.OAbstractWALRecord;
 import com.orientechnologies.orient.core.storage.impl.local.paginated.wal.OLogSequenceNumber;
@@ -38,42 +36,40 @@ import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Locale;
 import java.util.Random;
 
 public class ReadWriteDiskCacheTest {
-  private static final int userDataSize            = 8;
-  private static final int writeCacheAmountOfPages = 15000;
-  private static final int systemOffset            = OIntegerSerializer.INT_SIZE + 3 * OLongSerializer.LONG_SIZE;
-  private static final int PAGE_SIZE               = userDataSize + systemOffset;
-  private static final int READ_CACHE_MAX_MEMORY   = 4 * PAGE_SIZE;
-  private static final int WRITE_CACHE_MAX_SIZE    = writeCacheAmountOfPages * PAGE_SIZE;
-  private static O2QCache  readBuffer;
-  private static OWOWCache writeBuffer;
+  private static final int                   userDataSize            = 8;
+  private static final int                   writeCacheAmountOfPages = 15000;
+  private static final int                   systemOffset            = OIntegerSerializer.INT_SIZE + 3 * OLongSerializer.LONG_SIZE;
+  private static final int                   PAGE_SIZE               = userDataSize + systemOffset;
+  private static final int                   READ_CACHE_MAX_MEMORY   = 4 * PAGE_SIZE;
+  private static final int                   WRITE_CACHE_MAX_SIZE    = writeCacheAmountOfPages * PAGE_SIZE;
+  private static       O2QCache              readBuffer;
+  private static       OWOWCache             writeBuffer;
+  private static       OCASDiskWriteAheadLog writeAheadLog;
+
   private static final OClosableLinkedContainer<Long, OFileClassic> files = new OClosableLinkedContainer<>(1024);
 
-  private static OLocalPaginatedStorage storageLocal;
-  private static String                 fileName;
-  private static String                 storagePath;
-  private static OCASDiskWriteAheadLog  writeAheadLog;
-  private        byte                   seed;
+  private static String fileName;
+  private static Path   storagePath;
+  private static String storageName;
+  private        byte   seed;
 
   private static final OByteBufferPool BUFFER_POOL = new OByteBufferPool(PAGE_SIZE);
 
   @BeforeClass
-  public static void beforeClass() throws IOException {
-    OGlobalConfiguration.FILE_LOCK.setValue(Boolean.FALSE);
+  public static void beforeClass() {
     OGlobalConfiguration.STORAGE_EXCLUSIVE_FILE_ACCESS.setValue(Boolean.FALSE);
 
     String buildDirectory = System.getProperty("buildDirectory");
-    if (buildDirectory == null)
+    if (buildDirectory == null) {
       buildDirectory = ".";
+    }
 
-    storagePath = buildDirectory + "/ReadWriteDiskCacheTest";
-    storageLocal = (OLocalPaginatedStorage) Orient.instance().getRunningEngine("plocal")
-        .createStorage(storagePath, null, 125 * 1024 * 1024);
-    storageLocal.create(new OContextConfiguration());
-    storageLocal.close(true, false);
-
+    storageName = "ReadWriteDiskCacheTest";
+    storagePath = Paths.get(buildDirectory).resolve("ReadWriteDiskCacheTest");
     fileName = "readWriteDiskCacheTest.tst";
 
     OWALRecordsFactory.INSTANCE.registerNewRecord((byte) 128, TestRecord.class);
@@ -95,10 +91,12 @@ public class ReadWriteDiskCacheTest {
     }
 
     if (writeBuffer != null) {
-      if (readBuffer != null)
+      if (readBuffer != null) {
         readBuffer.deleteStorage(writeBuffer);
-      else
+      } else {
         writeBuffer.delete();
+      }
+
       writeBuffer = null;
     }
 
@@ -112,10 +110,8 @@ public class ReadWriteDiskCacheTest {
       writeAheadLog = null;
     }
 
-    storageLocal.delete();
-
     if (nativeFileName != null) {
-      File file = new File(storageLocal.getConfiguration().getDirectory() + "/" + nativeFileName);
+      final File file = storagePath.resolve(nativeFileName).toFile();
 
       if (file.exists()) {
         Assert.assertTrue(file.delete());
@@ -124,7 +120,6 @@ public class ReadWriteDiskCacheTest {
     }
 
     BUFFER_POOL.clear();
-    OGlobalConfiguration.FILE_LOCK.setValue(Boolean.TRUE);
     OGlobalConfiguration.STORAGE_EXCLUSIVE_FILE_ACCESS.setValue(Boolean.TRUE);
   }
 
@@ -152,43 +147,55 @@ public class ReadWriteDiskCacheTest {
     }
 
     if (writeBuffer != null) {
-      if (readBuffer != null)
-        readBuffer.closeStorage(writeBuffer);
-      else
-        writeBuffer.close();
+      if (readBuffer != null) {
+        readBuffer.deleteStorage(writeBuffer);
+      } else {
+        writeBuffer.delete();
+      }
+
       writeBuffer = null;
     }
 
     if (readBuffer != null) {
       readBuffer.clear();
+
       readBuffer = null;
     }
 
     if (writeAheadLog != null) {
       writeAheadLog.delete();
+
       writeAheadLog = null;
     }
+
     files.clear();
 
-    File testFile = new File(storageLocal.getConfiguration().getDirectory() + File.separator + nativeFileName);
-    if (testFile.exists()) {
-      Assert.assertTrue(testFile.delete());
+    if (nativeFileName != null) {
+      final File testFile = storagePath.resolve(nativeFileName).toFile();
+
+      if (testFile.exists()) {
+        Assert.assertTrue(testFile.delete());
+      }
     }
 
-    File idMapFile = new File(storageLocal.getConfiguration().getDirectory() + File.separator + "name_id_map.cm");
+    File idMapFile = storagePath.resolve("name_id_map.cm").toFile();
     if (idMapFile.exists()) {
       Assert.assertTrue(idMapFile.delete());
     }
 
-    idMapFile = new File(storageLocal.getConfiguration().getDirectory() + File.separator + "name_id_map_v2.cm");
+    idMapFile = storagePath.resolve("name_id_map_v2.cm").toFile();
     if (idMapFile.exists()) {
       Assert.assertTrue(idMapFile.delete());
     }
   }
 
   private void initBuffer() throws IOException, InterruptedException {
-    writeBuffer = new OWOWCache(PAGE_SIZE, BUFFER_POOL, writeAheadLog, -1, WRITE_CACHE_MAX_SIZE, storageLocal, false, files, 1,
-        OChecksumMode.StoreAndThrow);
+    writeAheadLog = new OCASDiskWriteAheadLog(storageName, storagePath, storagePath, 12_000, Integer.MAX_VALUE, Integer.MAX_VALUE,
+        25, true, Locale.US, -1, 1024L * 1024 * 1024, 1000, true, false, true, 10);
+
+    writeBuffer = new OWOWCache(PAGE_SIZE, BUFFER_POOL, writeAheadLog, -1, WRITE_CACHE_MAX_SIZE, READ_CACHE_MAX_MEMORY, storagePath,
+        storageName, OStringSerializer.INSTANCE, files, 1, OChecksumMode.StoreAndThrow, false, false, 0.9, true, 10, true, true,
+        true);
     writeBuffer.loadRegisteredFiles();
 
     readBuffer = new O2QCache(READ_CACHE_MAX_MEMORY, PAGE_SIZE, false, 50, true, 10);
@@ -207,8 +214,6 @@ public class ReadWriteDiskCacheTest {
         entries[i] = readBuffer.allocateNewPage(fileId, writeBuffer, true);
         Assert.assertEquals(entries[i].getPageIndex(), i);
       }
-
-      entries[i].markDirty();
 
       final ByteBuffer buffer = entries[i].getCachePointer().getBuffer();
 
@@ -278,8 +283,6 @@ public class ReadWriteDiskCacheTest {
 
         entries[i].getCachePointer().acquireExclusiveLock();
 
-        entries[i].markDirty();
-
         final ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
 
         buffer.position(systemOffset);
@@ -311,8 +314,7 @@ public class ReadWriteDiskCacheTest {
       for (int i = 3; i >= 0; i--) {
         final String fileNativeName = nativeFileNames.get(i);
 
-        String path = storageLocal.getConfiguration().getDirectory() + "/" + fileNativeName;
-        final File file = new File(path);
+        final File file = storagePath.resolve(fileNativeName).toFile();
         if (i > n) {
           Assert.assertFalse(file.exists());
         } else {
@@ -325,8 +327,7 @@ public class ReadWriteDiskCacheTest {
       for (int i = 3; i >= 0; i--) {
         final String fileNativeName = nativeFileNames.get(i);
 
-        String path = storageLocal.getConfiguration().getDirectory() + "/" + fileNativeName;
-        final File file = new File(path);
+        final File file = storagePath.resolve(fileNativeName).toFile();
         if (i >= n) {
           Assert.assertFalse(file.exists());
         } else {
@@ -367,8 +368,6 @@ public class ReadWriteDiskCacheTest {
 
         entries[i].getCachePointer().acquireExclusiveLock();
 
-        entries[i].markDirty();
-
         final ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
 
         buffer.position(systemOffset);
@@ -381,6 +380,7 @@ public class ReadWriteDiskCacheTest {
     }
 
     readBuffer.closeStorage(writeBuffer);
+    writeAheadLog.close();
 
     initBuffer();
 
@@ -402,8 +402,7 @@ public class ReadWriteDiskCacheTest {
       for (int i = 3; i >= 0; i--) {
         final String fileNativeName = nativeFileNames.get(i);
 
-        String path = storageLocal.getConfiguration().getDirectory() + "/" + fileNativeName;
-        final File file = new File(path);
+        final File file = storagePath.resolve(fileNativeName).toFile();
         if (i > n) {
           Assert.assertFalse(file.exists());
         } else {
@@ -416,8 +415,7 @@ public class ReadWriteDiskCacheTest {
       for (int i = 3; i >= 0; i--) {
         final String fileNativeName = nativeFileNames.get(i);
 
-        String path = storageLocal.getConfiguration().getDirectory() + "/" + fileNativeName;
-        final File file = new File(path);
+        final File file = storagePath.resolve(fileNativeName).toFile();
         if (i >= n) {
           Assert.assertFalse(file.exists());
         } else {
@@ -441,8 +439,6 @@ public class ReadWriteDiskCacheTest {
         entries[i] = readBuffer.allocateNewPage(fileId, writeBuffer, true);
         Assert.assertEquals(entries[i].getPageIndex(), i);
       }
-
-      entries[i].markDirty();
 
       final ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
       buffer.position(systemOffset);
@@ -499,8 +495,7 @@ public class ReadWriteDiskCacheTest {
     final long fileId = readBuffer.addFile(fileName, writeBuffer);
     final String nativeFileName = writeBuffer.nativeFileNameById(fileId);
 
-    File file = new File(storageLocal.getConfiguration().getDirectory() + "/" + nativeFileName);
-
+    File file = storagePath.resolve(nativeFileName).toFile();
     Assert.assertTrue(file.exists());
     Assert.assertTrue(file.isFile());
   }
@@ -518,8 +513,6 @@ public class ReadWriteDiskCacheTest {
         entries[i] = readBuffer.allocateNewPage(fileId, writeBuffer, true);
         Assert.assertEquals(entries[i].getPageIndex(), i);
       }
-
-      entries[i].markDirty();
 
       final ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
       buffer.position(systemOffset);
@@ -593,8 +586,6 @@ public class ReadWriteDiskCacheTest {
         Assert.assertEquals(entries[i].getPageIndex(), i);
       }
 
-      entries[i].markDirty();
-
       final ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
       buffer.position(systemOffset);
       buffer.put(new byte[] { (byte) i, 1, 2, seed, 4, 5, 6, (byte) i });
@@ -649,8 +640,6 @@ public class ReadWriteDiskCacheTest {
         buffer.put(userData);
 
         setLsn(buffer, new OLogSequenceNumber(1, i));
-
-        cacheEntry.markDirty();
       } finally {
         readBuffer.releaseFromWrite(cacheEntry, writeBuffer);
       }
@@ -717,8 +706,6 @@ public class ReadWriteDiskCacheTest {
         buffer.put(userData);
 
         setLsn(buffer, new OLogSequenceNumber(1, i));
-
-        cacheEntry.markDirty();
       } finally {
         readBuffer.releaseFromWrite(cacheEntry, writeBuffer);
       }
@@ -803,8 +790,6 @@ public class ReadWriteDiskCacheTest {
         buffer.put(userData);
 
         setLsn(buffer, new OLogSequenceNumber(1, i));
-
-        cacheEntry.markDirty();
       } finally {
         readBuffer.releaseFromWrite(cacheEntry, writeBuffer);
       }
@@ -849,7 +834,7 @@ public class ReadWriteDiskCacheTest {
 
     //move page 4 to pinned pages
     cacheEntry = readBuffer.loadForRead(fileId, 4, false, writeBuffer, 1, true);
-    readBuffer.pinPage(cacheEntry);
+    readBuffer.pinPage(cacheEntry, writeBuffer);
     readBuffer.releaseFromRead(cacheEntry, writeBuffer);
 
     Assert.assertEquals(am.size(), 1); //page 1
@@ -924,7 +909,6 @@ public class ReadWriteDiskCacheTest {
         Assert.assertEquals(entries[i].getPageIndex(), i);
       }
 
-      entries[i].markDirty();
       ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
       buffer.position(systemOffset);
       buffer.put(new byte[] { (byte) i, 1, 2, seed, 4, 5, 6, (byte) i });
@@ -981,7 +965,7 @@ public class ReadWriteDiskCacheTest {
     writeBuffer.flush();
 
     for (int i = 0; i < 4; i++) {
-      File file = new File(storageLocal.getConfiguration().getDirectory() + "/" + nativeFileName);
+      File file = storagePath.resolve(nativeFileName).toFile();
       Assert.assertFalse(file.exists());
     }
   }
@@ -992,7 +976,7 @@ public class ReadWriteDiskCacheTest {
 
     final long fileId = writeBuffer.addFile(fileName);
     final String nativeFileName = writeBuffer.nativeFileNameById(fileId);
-    final Path path = Paths.get(storageLocal.getConfiguration().getDirectory(), nativeFileName);
+    final Path path = storagePath.resolve(nativeFileName);
     Assert.assertTrue(Files.exists(path));
 
     // Set the file content to random.
@@ -1054,8 +1038,6 @@ public class ReadWriteDiskCacheTest {
           Assert.assertEquals(entries[i].getPageIndex(), i);
         }
 
-        entries[i].markDirty();
-
         final ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
         buffer.position(systemOffset);
         buffer.put(new byte[] { (byte) i, 1, 2, seed, 4, 5, (byte) j, (byte) i });
@@ -1102,8 +1084,6 @@ public class ReadWriteDiskCacheTest {
         Assert.assertEquals(entries[i].getPageIndex(), i);
       }
 
-      entries[i].markDirty();
-
       final ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
       buffer.position(systemOffset);
       buffer.put(new byte[] { (byte) i, 1, 2, seed, 4, 5, 6, 7 });
@@ -1149,8 +1129,6 @@ public class ReadWriteDiskCacheTest {
           Assert.assertEquals(entries[i].getPageIndex(), i);
         }
 
-        entries[i].markDirty();
-
         ByteBuffer buffer = entries[i].getCachePointer().getBufferDuplicate();
         buffer.position(systemOffset);
         buffer.put(new byte[] { (byte) i, 1, 2, seed, 4, 5, 6, 7 });
@@ -1165,18 +1143,19 @@ public class ReadWriteDiskCacheTest {
       }
     } finally {
       for (int i = 0; i < 4; i++) {
+        assert entries[i] != null;
         readBuffer.releaseFromWrite(entries[i], writeBuffer);
       }
     }
   }
 
   private void assertFile(long pageIndex, byte[] value, OLogSequenceNumber lsn, String fileName) throws IOException {
-    String path = storageLocal.getConfiguration().getDirectory() + "/" + fileName;
-
-    OFileClassic fileClassic = new OFileClassic(Paths.get(path));
+    OFileClassic fileClassic = new OFileClassic(storagePath.resolve(fileName));
     fileClassic.open();
     byte[] content = new byte[userDataSize + systemOffset];
-    fileClassic.read(pageIndex * (userDataSize + systemOffset), content, userDataSize + systemOffset);
+    final ByteBuffer byteBuffer = ByteBuffer.wrap(content);
+
+    fileClassic.read(pageIndex * (userDataSize + systemOffset), byteBuffer, true);
 
     Assertions.assertThat(Arrays.copyOfRange(content, systemOffset, userDataSize + systemOffset)).isEqualTo(value);
